added descriptions and status responses for all endpoints

+++ app/endpoints/api_garden.py
@@ -10,6 +10,7 @@ from app.database.session import get_session
 from app.library.helpers import *
 from app.models.garden_models import Bed, BedCreate, BedRead, BedUpdate
 from app.models.garden_models import Planting, PlantingCreate, PlantingRead, PlantingUpdate
+from app.models.user_models import User
 from app.endpoints.api_user import auth_handler
 
 garden_router = APIRouter()
@@ -17,13 +18,14 @@ garden_router = APIRouter()
 
 # CRUD API methods for Garden Beds
 
-@garden_router.post("/api/beds/", response_model=BedRead, tags=["Garden Beds API"])
+@garden_router.post("/api/beds/", status_code=status.HTTP_201_CREATED, response_model=BedRead, tags=["Garden Beds API"])
 def create_bed(*,
                session: Session = Depends(get_session),
                response: Response,
-               user=Depends(auth_handler.get_current_user),
+               user: User = Depends(auth_handler.get_current_user),
                bed: BedCreate
                ):
+  """Create a garden bed."""
   if not user.gardener:
     response.status_code = status.HTTP_401_UNAUTHORIZED
     return {}
@@ -41,7 +43,7 @@ def read_beds(*,
               offset: int = 0,
               limit: int = Query(default=100, lte=100)
               ):
-  # select * from
+  """Get the list of defined garden beds."""
   stmt = select(Bed).offset(offset).limit(limit)
   db_beds = session.exec(stmt).all()
   return db_beds
@@ -49,22 +51,28 @@ def read_beds(*,
 
 @garden_router.get("/api/beds/{bed_id}", response_model=BedRead, tags=["Garden Beds API"])
 def read_bed(*, session: Session = Depends(get_session), bed_id: int):
-  # find the planting with the given ID, or None if it does not exist
+  """Get the garden bed with the given ID, or None if it does not exist."""
   db_bed = session.get(Bed, bed_id)
   if not db_bed:
-    raise HTTPException(status_code=404, detail='Bed not found')
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Bed not found')
   return db_bed
 
 
-@garden_router.patch("/api/beds/{bed_id}", response_model=BedRead, tags=["Garden Beds API"])
+@garden_router.patch("/api/beds/{bed_id}", status_code=status.HTTP_201_CREATED, response_model=BedRead, tags=["Garden Beds API"])
 def update_bed(*,
                session: Session = Depends(get_session),
+               response: Response,
+               user: User = Depends(auth_handler.get_current_user),
                bed_id: int,
                bed: BedUpdate,
                ):
+  """Update the details of the garden bed with the given ID."""
+  if not user.gardener:
+    response.status_code = status.HTTP_401_UNAUTHORIZED
+    return {}
   db_bed = session.get(Bed, bed_id)
   if not db_bed:
-    raise HTTPException(status_code=404, detail='Bed not found')
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Bed not found')
   # update the planting data
   bed_data = bed.dict(exclude_unset=True)
   for key, val in bed_data.items():
@@ -75,14 +83,20 @@ def update_bed(*,
   return db_bed
 
 
-@garden_router.delete("/api/beds/{bed_id}", tags=["Garden Beds API"])
+@garden_router.delete("/api/beds/{bed_id}", status_code=status.HTTP_202_ACCEPTED, tags=["Garden Beds API"])
 def delete_bed(*,
                session: Session = Depends(get_session),
+               response: Response,
+               user: User = Depends(auth_handler.get_current_user),
                bed_id: int,
                ):
+  """Delete the garden bed with the given ID."""
+  if not user.gardener:
+    response.status_code = status.HTTP_401_UNAUTHORIZED
+    return {}
   db_bed = session.get(Bed, bed_id)
   if not db_bed:
-    raise HTTPException(status_code=404, detail='Bed not found')
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Bed not found')
   session.delete(db_bed)
   session.commit()
   return {'ok': True}
@@ -90,11 +104,12 @@ def delete_bed(*,
 
 # CRUD API methods for Garden Plantings
 
-@garden_router.post("/api/plantings/", response_model=PlantingRead, tags=["Garden Plantings API"])
+@garden_router.post("/api/plantings/", response_model=PlantingRead, status_code=status.HTTP_201_CREATED, tags=["Garden Plantings API"])
 def create_planting(*,
                     session: Session = Depends(get_session),
                     planting: PlantingCreate
                     ):
+  """Create a garden planting."""
   db_planting = Planting.from_orm(planting)
   session.add(db_planting)
   session.commit()
@@ -108,7 +123,7 @@ def read_plantings(*,
                    offset: int = 0,
                    limit: int = Query(default=100, lte=100)
                    ):
-  # select * from
+  """Get the list of defined garden plantings."""
   stmt = select(Planting).offset(offset).limit(limit)
   db_plantings = session.exec(stmt).all()
   return db_plantings
@@ -118,19 +133,20 @@ def read_plantings(*,
 def read_planting(*,
                   session: Session = Depends(get_session),
                   planting_id: int = Path(None, description="The ID of the planting  to return")):
-  # find the planting with the given ID, or None if it does not exist
+  """Get the garden planting with the given ID, or None if it does not exist."""
   db_planting = session.get(Planting, planting_id)
   if not db_planting:
     raise HTTPException(status_code=404, detail="Planting not found")
   return db_planting
 
 
-@garden_router.patch("/api/plantings/{planting_id}", response_model=PlantingRead, tags=["Garden Plantings API"])
+@garden_router.patch("/api/plantings/{planting_id}", response_model=PlantingRead, status_code=status.HTTP_201_CREATED, tags=["Garden Plantings API"])
 def update_planting(*,
                     session: Session = Depends(get_session),
                     planting_id: int,
                     planting: PlantingUpdate,
                     ):
+  """Update the details of the garden planting with the given ID."""
   db_planting = session.get(Planting, planting_id)
   if not db_planting:
     raise HTTPException(status_code=404, detail="Planting not found")
@@ -144,11 +160,12 @@ def update_planting(*,
   return db_planting
 
 
-@garden_router.delete("/api/plantings/{planting_id}", tags=["Garden Plantings API"])
+@garden_router.delete("/api/plantings/{planting_id}", status_code=status.HTTP_202_ACCEPTED, tags=["Garden Plantings API"])
 def delete_planting(*,
                     session: Session = Depends(get_session),
                     planting_id: int,
                     ):
+  """Delete the garden planting with the given ID."""
   db_planting = session.get(Planting, planting_id)
   if not db_planting:
     raise HTTPException(status_code=404, detail="Planting not found")
